+++ apps/web/src/app/responsible/route.ts
@@ -1,8 +1,0 @@
-import { responsiblePage } from "../../../../../apps/leaderboard/src/pages/responsible.js";
-import { marketingResponse } from "@/lib/marketing";
-
-export const dynamic = "force-dynamic";
-
-export async function GET(): Promise<Response> {
-  return marketingResponse(responsiblePage);
-}